style admin details panel
