+++ frontend/src/components/DetailsPanel/index.css
@@ -31,7 +31,7 @@
      border-bottom: 1px solid #e3e3e3;
      display: flex;
      flex-direction: column;
-     gap: 4px;
+     gap: 20px;
 }
 
 .panel-header-date-details {
@@ -41,6 +41,12 @@
      display: flex;
      flex-direction: column;
      gap: 4px;
+}
+
+.panel-class-schedules {
+     display: flex;
+     flex-direction: column;
+     gap: 6px;
 }
 
 .panel-header-class-name {
@@ -92,7 +98,7 @@
 .panel-details-shift {
      display: flex;
      flex-direction: column;
-     gap: 12px;
+     gap: 16px;
 }
 
 .panel-details-shift-row {
@@ -121,11 +127,8 @@
      margin-right: 6px;
 }
 
-.panel-details-description {
-     margin-top: 20px;
-}
-
 .panel-description {
+     margin-top: 10px;
      margin-top: 0.5rem;
      font-family: 'Roboto';
      font-weight: 400;
@@ -133,88 +136,17 @@
      line-height: 21.6px;
 }
 
-.button-icons {
-     display: flex;
-     flex-direction: row;
-     gap: 4px;
-     width: fit-content;
-     margin-left: auto;
-     margin-right: 36px;
-     margin-top: 20px;
-}
-
- /* Shift Details Panel */
-/* Dynamic Button Styling based on Shift Type */
-.panel-buttons {
-     margin-top: 4vh;
-     height: 4.5vh;
-     display: flex;
-     flex-direction: row;
-     gap: 12px;
-}
-
-.dynamic-button {
-     position: relative;
-     height: 100%;
-     border-radius: 8px;
-     border: 1px solid var(--primary-blue);
-     background: #FFF;
-     box-shadow: 0px 4px 16px 0px rgba(15, 17, 17, 0.05);
-     padding: 0 0.75rem;
-     display: flex;
-     align-items: center;
- }
- 
- .dynamic-button:not(:disabled):hover {
-     cursor: pointer;
-     background: var(--button-blue);
-     border: none;
- }
- 
- .dynamic-button:disabled {
-     background-color: #d3d3d3;
-     border: 1px solid #d3d3d3;
-     box-shadow: none;
- }
-
- .dynamic-button.primary-action {
-     background-color: var(--button-blue);
-     color: var(--white);
-     border: var(--button-blue);
- }
-
- .dynamic-button.primary-action:hover {
-     background-color: var(--button-blue-hover);
- }
-
- .dynamic-button.cancel-action {
-     background-color: var(--red);
-     color: var(--white);
-     border: var(--red);
- }
-
- .dynamic-button.cancel-action:hover {
-     background-color: var(--red-hover);
- }
- 
- /* Icon within Button */
- .card-button-icon {
-     width: 1rem;
-     height: 1rem;
-     margin-right: .5rem;
- }
-
- .zoom-icon {
+.zoom-icon {
      height: 1.5rem;
      width: 1.5rem;
      margin-right: 0.5rem; 
- }
+}
 
- .email-icon {
+.email-icon {
      margin-left: 0.5rem;
- }
+}
  
- .join-class-button {
+.join-class-button {
      display: flex; 
      flex-direction: row; 
      align-items: center; 
@@ -225,23 +157,23 @@
      background: #FFF;
      cursor: pointer;
      transition: background-color 0.3s ease;
- }
+}
  
- .join-class-button a {
+.join-class-button a {
      text-decoration: none;
      color: var(--dark);
      display: flex; 
      flex-direction: row;
      align-items: center; 
- }
+}
  
- .join-class-button:hover {
+.join-class-button:hover {
      cursor: pointer;
      background: var(--button-blue);
- }
+}
 
- .class-volunteers {
+.class-volunteers {
      display: flex;
      flex-direction: column;
      gap: 8px;
- }
+}
