+++ frontend/src/components/VolunteerDetailsPanel/index.css
@@ -64,22 +64,8 @@
 .volunteer-profile {
      width: 28px;
      height: 28px;
-     background-color: #808080;
      border-radius: 4px;
-     opacity: 0.3;
-     margin-right: 6px;
-}
-
-.panel-details-description {
-     margin-top: 20px;
-}
-
-.panel-description {
-     margin-top: 10px;
-     font-family: 'Roboto';
-     font-weight: 400;
-     font-size: 16px;
-     line-height: 21.6px;
+     margin-right: 10px;
 }
 
 .button-icons {
@@ -89,14 +75,13 @@
      width: fit-content;
      margin-left: auto;
      margin-right: 36px;
-     margin-top: 20px;
 }
 
- /* Shift Details Panel */
+/* Shift Details Panel */
 /* Dynamic Button Styling based on Shift Type */
 .panel-buttons {
-     padding-top: 2vh;
-     height: 4vh;
+     margin-top: 4vh;
+     height: 4.5vh;
      display: flex;
      flex-direction: row;
      gap: 12px;
@@ -109,36 +94,46 @@
      border: 1px solid var(--primary-blue);
      background: #FFF;
      box-shadow: 0px 4px 16px 0px rgba(15, 17, 17, 0.05);
-     padding: 0 1rem;
+     padding: 0 0.75rem;
      display: flex;
      align-items: center;
  }
  
- .dynamic-button:not(:disabled):hover {
+.dynamic-button:not(:disabled):hover {
      cursor: pointer;
      background: var(--button-blue);
      border: none;
- }
+}
  
- .dynamic-button:disabled {
+.dynamic-button:disabled {
      background-color: #d3d3d3;
      border: 1px solid #d3d3d3;
      box-shadow: none;
- }
+}
+
+.dynamic-button.primary-action {
+     background-color: var(--button-blue);
+     color: var(--white);
+     border: var(--button-blue);
+}
+
+.dynamic-button.primary-action:hover {
+     background-color: var(--button-blue-hover);
+}
+
+.dynamic-button.cancel-action {
+     background-color: var(--red);
+     color: var(--white);
+     border: var(--red);
+}
+
+.dynamic-button.cancel-action:hover {
+     background-color: var(--red-hover);
+}
  
  /* Icon within Button */
- .card-button-icon {
+.card-button-icon {
      width: 1rem;
      height: 1rem;
      margin-right: .5rem;
- }
-
- .zoom-icon {
-     height: 1.5rem;
-     width: 1.5rem;
-     margin-right: 0.5rem; 
- }
-
- .email-icon {
-     margin-left: 0.5rem;
- }
+}
